+++ src/cobib/database/__init__.py
@@ -4,10 +4,6 @@ coBib uses a plain-test YAML file to store the bibliographic database.
 This module contains the components which represent this database at runtime.
 """
 
-from .database import Database
-from .entry import Entry
-
-__all__ = [
-    "Database",
-    "Entry",
-]
+from .author import Author as Author
+from .database import Database as Database
+from .entry import Entry as Entry